Rename the project from vaadin-services to vaadin-connect
