+++ README.md
@@ -1,2 +1,2 @@
-# vaadin-services
+# vaadin-connect
 A Vaadin Labs experiment with a secure stateless communication framework
