+++ pom.xml
@@ -7,16 +7,16 @@
         <version>2.0.0</version>
         <relativePath />
     </parent>
-    <artifactId>vaadin-services-parent</artifactId>
+    <artifactId>vaadin-connect-parent</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>Vaadin Services Parent</name>
-    <description>Vaadin Services Parent</description>
+    <name>Vaadin Connect Parent</name>
+    <description>Vaadin Connect Parent</description>
     <url>http://vaadin.com</url>
 
     <modules>
-        <module>vaadin-services</module>
-        <module>vaadin-services-demo</module>
+        <module>vaadin-connect</module>
+        <module>vaadin-connect-demo</module>
     </modules>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -27,9 +27,9 @@
         <jacoco.version>0.8.2</jacoco.version>
     </properties>
     <scm>
-        <url>https://github.com/vaadin/vaadin-services</url>
-        <connection>scm:git:https://github.com/vaadin/vaadin-services.git</connection>
-        <developerConnection>scm:git:git@github.com:vaadin/vaadin-services.git</developerConnection>
+        <url>https://github.com/vaadin/vaadin-connect</url>
+        <connection>scm:git:https://github.com/vaadin/vaadin-connect.git</connection>
+        <developerConnection>scm:git:git@github.com:vaadin/vaadin-connect.git</developerConnection>
     </scm>
     <dependencyManagement>
         <dependencies>
