+++ vaadin-connect-demo/src/main/java/com/vaadin/connect/VaadinConnectDemoApplication.java
@@ -13,14 +13,24 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.services;
+package com.vaadin.connect;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
- * Dummy class to verify project setup works.
+ * Main class of the Vaadin connect demo module.
  */
-public class DummyClass {
+@SpringBootApplication
+public class VaadinConnectDemoApplication {
 
-  protected boolean getTrue() {
-    return true;
+  /**
+   * Main method to run the application.
+   *
+   * @param args
+   *          arguments
+   */
+  public static void main(String[] args) {
+    SpringApplication.run(VaadinConnectDemoApplication.class, args);
   }
 }
